Rename: 게시글 작성 페이지 폴더 변경 및 경로 수정 (#81)
